# core-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRAWOBIORCA",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/regulations": {
      "get": {
        "tags": [
          "regulations"
        ],
        "summary": "Get Public Regulations",
        "operationId": "get_public_regulations_api_regulations_get",
        "parameters": [
          {
            "name": "documentType",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RegulationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Documenttype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegulationRepresentation"
                  },
                  "title": "Response Get Public Regulations Api Regulations Get"
                }
              }
            }
          },
          "204": {
            "description": "No public files for given search criteria."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "regulations"
        ],
        "summary": "Add Public Regulation",
        "operationId": "add_public_regulation_api_regulations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulationData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulationUploadTarget"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regulations/{regulationId}/documents": {
      "get": {
        "tags": [
          "regulations"
        ],
        "summary": "Search Regulation Documents",
        "operationId": "search_regulation_documents_api_regulations__regulationId__documents_get",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": -1,
              "title": "Threshold"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SearchOrder",
              "default": "document"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search Regulation Documents Api Regulations  Regulationid  Documents Get"
                }
              }
            }
          },
          "204": {
            "description": "No search results."
          },
          "400": {
            "description": "Regulation not prepared, normally should not occur."
          },
          "404": {
            "description": "Regulation not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regulations/{regulationId}/confirm-upload": {
      "post": {
        "tags": [
          "regulations"
        ],
        "summary": "Confirm Public Regulation Upload",
        "operationId": "confirm_public_regulation_upload_api_regulations__regulationId__confirm_upload_post",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Regulation not found!"
          },
          "409": {
            "description": "Regulation upload already confirmed, or its content not found on storage!"
          },
          "503": {
            "description": "Preparation service not working!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regulations/{regulationId}/download-url": {
      "get": {
        "tags": [
          "regulations"
        ],
        "summary": "Get Public Regulation Download Url",
        "operationId": "get_public_regulation_download_url_api_regulations__regulationId__download_url_get",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Public Regulation Download Url Api Regulations  Regulationid  Download Url Get"
                }
              }
            }
          },
          "404": {
            "description": "Regulation not found!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regulations/{regulationId}": {
      "delete": {
        "tags": [
          "regulations"
        ],
        "summary": "Delete Public Regulation",
        "operationId": "delete_public_regulation_api_regulations__regulationId__delete",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Regulation not found!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regulations/{regulationId}/preparation-retry": {
      "post": {
        "tags": [
          "regulations"
        ],
        "summary": "Retry Public Regulation Preparation",
        "operationId": "retry_public_regulation_preparation_api_regulations__regulationId__preparation_retry_post",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Regulation to prepare not found!"
          },
          "409": {
            "description": "Regulation already prepared, preparation in progress, or its content not uploaded!"
          },
          "503": {
            "description": "Preparation service not working!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/regulations": {
      "get": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Get User Regulations",
        "operationId": "get_user_regulations_api_user_regulations_get",
        "parameters": [
          {
            "name": "documentType",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RegulationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Documenttype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegulationRepresentation"
                  },
                  "title": "Response Get User Regulations Api User Regulations Get"
                }
              }
            }
          },
          "204": {
            "description": "Not found user files with that criteria."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Add User Regulation",
        "operationId": "add_user_regulation_api_user_regulations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulationData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulationUploadTarget"
                }
              }
            }
          },
          "400": {
            "description": "Can't add empty regulation."
          },
          "404": {
            "description": "User not found!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/regulations/{regulationId}/confirm-upload": {
      "post": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Confirm User Regulation Upload",
        "operationId": "confirm_user_regulation_upload_api_user_regulations__regulationId__confirm_upload_post",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Regulation not found!"
          },
          "409": {
            "description": "Regulation upload already confirmed, or its content not found on storage!"
          },
          "503": {
            "description": "Preparation service not working!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/regulations/{regulationId}/download-url": {
      "get": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Get User Regulation Download Url",
        "operationId": "get_user_regulation_download_url_api_user_regulations__regulationId__download_url_get",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get User Regulation Download Url Api User Regulations  Regulationid  Download Url Get"
                }
              }
            }
          },
          "404": {
            "description": "Regulation not found!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/regulations/{regulationId}": {
      "delete": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Delete User Regulation",
        "operationId": "delete_user_regulation_api_user_regulations__regulationId__delete",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Regulation not found!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/regulations/{regulationId}/documents": {
      "get": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Search Regulation Documents",
        "operationId": "search_regulation_documents_api_user_regulations__regulationId__documents_get",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": -1,
              "title": "Threshold"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SearchOrder",
              "default": "document"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search Regulation Documents Api User Regulations  Regulationid  Documents Get"
                }
              }
            }
          },
          "204": {
            "description": "No search results."
          },
          "400": {
            "description": "Regulation not prepared, by user, can't search."
          },
          "404": {
            "description": "Regulation not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/regulations/{regulationId}/preparation-retry": {
      "post": {
        "tags": [
          "user_regulations"
        ],
        "summary": "Retry User Regulation Preparation",
        "operationId": "retry_user_regulation_preparation_api_user_regulations__regulationId__preparation_retry_post",
        "parameters": [
          {
            "name": "regulationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Regulationid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Regulation to prepare not found!"
          },
          "409": {
            "description": "Regulation already prepared, preparation in progress, or its content not uploaded!"
          },
          "503": {
            "description": "Preparation service not working!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegulationData": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "regulationType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegulationType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RegulationData"
      },
      "RegulationPreparationStatus": {
        "type": "string",
        "enum": [
          "NOT_STARTED",
          "IN_PROGRESS",
          "PREPARED",
          "FAILED"
        ],
        "title": "RegulationPreparationStatus"
      },
      "RegulationRepresentation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "presentationName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Presentationname"
          },
          "regulationType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegulationType"
              },
              {
                "type": "null"
              }
            ]
          },
          "preparationStatus": {
            "$ref": "#/components/schemas/RegulationPreparationStatus",
            "default": "NOT_STARTED"
          }
        },
        "type": "object",
        "required": [
          "id",
          "presentationName",
          "regulationType",
          "preparationStatus"
        ],
        "title": "RegulationRepresentation"
      },
      "RegulationType": {
        "type": "string",
        "enum": [
          "ACT",
          "DECREE",
          "STATUTE"
        ],
        "title": "RegulationType"
      },
      "RegulationUploadTarget": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "fields"
        ],
        "title": "RegulationUploadTarget"
      },
      "SearchOrder": {
        "type": "string",
        "enum": [
          "document",
          "score"
        ],
        "title": "SearchOrder"
      },
      "SearchResult": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Score"
          },
          "header": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Header"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "unit_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnitType"
              },
              {
                "type": "null"
              }
            ]
          },
          "unit_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Number"
          },
          "unit_path": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Path"
          },
          "elements": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SearchResultElement"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elements"
          },
          "highlight": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchResultHighlight"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "header",
          "text",
          "unit_type",
          "unit_number",
          "unit_path",
          "elements",
          "highlight"
        ],
        "title": "SearchResult"
      },
      "SearchResultElement": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "subsection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subsection"
          }
        },
        "type": "object",
        "required": [
          "text",
          "subsection"
        ],
        "title": "SearchResultElement"
      },
...
      "UnitType": {
        "type": "string",
        "enum": [
          "ARTICLE",
          "PARAGRAPH",
          "UNNUMBERED"
        ],
        "title": "UnitType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}